collect test repository implementations into TestHelpers module
