+++ test/project_types/script/commands/create_test.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require "project_types/script/test_helper"
-require "project_types/script/layers/infrastructure/fake_script_project_repository"
 
 module Script
   module Commands
@@ -18,7 +17,7 @@ module Script
         @script_name = "name"
         @ep_type = "discount"
         @no_config_ui = false
-        @script_project = Layers::Infrastructure::FakeScriptProjectRepository.new.create(
+        @script_project = TestHelpers::FakeScriptProjectRepository.new.create(
           language: @language,
           extension_point_type: @ep_type,
           script_name: @script_name,
