+++ test/project_types/script/layers/application/project_dependencies_test.rb
@@ -1,14 +1,14 @@
 # frozen_string_literal: true
 
 require "project_types/script/test_helper"
-require "project_types/script/layers/infrastructure/fake_extension_point_repository"
+require "project_types/script/test_helpers"
 
 describe Script::Layers::Application::ProjectDependencies do
   include TestHelpers::FakeFS
 
   let(:script_name) { "name" }
   let(:extension_point_type) { "discount" }
-  let(:extension_point_repository) { Script::Layers::Infrastructure::FakeExtensionPointRepository.new }
+  let(:extension_point_repository) { TestHelpers::FakeExtensionPointRepository.new }
   let(:extension_point) { extension_point_repository.get_extension_point(extension_point_type) }
   let(:task_runner) { stub }
