+++ test/project_types/script/test_helpers.rb
@@ -1,3 +1,5 @@
 module TestHelpers
-  autoload :FakeScriptProject, "project_types/script/test_helpers/fake_script_project"
+  autoload :FakeExtensionPointRepository, "project_types/script/test_helpers/fake_extension_point_repository"
+  autoload :FakePushPackageRepository, "project_types/script/test_helpers/fake_push_package_repository"
+  autoload :FakeScriptProjectRepository, "project_types/script/test_helpers/fake_script_project_repository"
 end
